+++ components/bridge/connection/orchestrator/impl/src/main/kotlin/com/flipperdevices/bridge/connection/ble/impl/FDeviceHolder.kt
@@ -7,6 +7,7 @@ import com.flipperdevices.bridge.connection.transport.common.api.FTransportConne
 import com.flipperdevices.core.ktx.jre.FlipperDispatchers
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
+import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.cancel
@@ -21,13 +22,15 @@ class FDeviceHolderFactory @Inject constructor(
     fun <API : FConnectedDeviceApi> build(
         config: FDeviceConnectionConfig<API>,
         listener: FTransportConnectionStatusListener,
-        onConnectError: (Throwable) -> Unit
+        onConnectError: (Throwable) -> Unit,
+        exceptionHandler: CoroutineExceptionHandler
     ): FDeviceHolder<API> {
         return FDeviceHolder(
             config = config,
             listener = listener,
             onConnectError = onConnectError,
-            deviceConnectionHelper = deviceConnectionHelper
+            deviceConnectionHelper = deviceConnectionHelper,
+            exceptionHandler = exceptionHandler
         )
     }
 }
@@ -35,11 +38,14 @@ class FDeviceHolder<API : FConnectedDeviceApi>(
     private val config: FDeviceConnectionConfig<API>,
     private val listener: FTransportConnectionStatusListener,
     private val onConnectError: (Throwable) -> Unit,
-    private val deviceConnectionHelper: FDeviceConfigToConnection
+    private val deviceConnectionHelper: FDeviceConfigToConnection,
+    private val exceptionHandler: CoroutineExceptionHandler
 ) : LogTagProvider {
     override val TAG = "FDeviceHolder-$config"
 
-    private val scope = CoroutineScope(FlipperDispatchers.workStealingDispatcher)
+    private val scope = CoroutineScope(
+        FlipperDispatchers.workStealingDispatcher + exceptionHandler
+    )
     private var deviceApi: API? = null
     private val connectJob: Job = scope.launch {
         deviceApi = deviceConnectionHelper.connect(